Change to count down and sundries
This changes the timer to count down instead of up. I also ran yarn format and made a couple linter corrections. Link was being imported but not used in the NotFound componoent, Countdown didn't have props validation, and the setInterval in Timer was returning an assignment which is no bueno.