Introduce Time util and TimeSets
This change creates the notion of a TimeSet, which is meant be how the user selects a time for the timer but may become more. It could have a color theme attached or something, users will eventually make custom TimeSets, etc. When making the TimeSetButton, I noticed that I needed to be able to display time just like i do in the Countdown, so I moved those time methods into a Time util file and added some tests for it. This also adds the TimeSetList and adds one to the Home page.